Test: Fixture for path to example GEFF file (#98)
# Proposed Change
Add pytest fixture for a path to example data and the expected graph
attributes. This can be used as an initial first pass test for new
backends, an example test for the networkx API has been added.

# Types of Changes
- Tests

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the developer/contributing docs.
- [x] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [x] I have checked that I maintained or improved code coverage.
- [ ] I have written docstrings and checked that they render correctly.
- [ ] If I changed the specification, I have checked that any validation
functions and tests reflect the changes.

---------

Co-authored-by: melisande-c <melisande.croft@fht.org>
Co-authored-by: Morgan Schwartz <msschwartz21@gmail.com>