102 nest geff metadata (#128)
# Proposed Change
Nest the geff metadata under the "geff" key instead of in the main
attrs.

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- New feature or enhancement
- Documentation update
- Tests

Which topics does your change affect? Delete those that do not apply.
- Specification
- `networkx` implementation

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the [developer/contributing](../CONTRIBUTING) docs.
- [x] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [x] I have checked that I maintained or improved code coverage.
- [x] I have written docstrings and checked that they render correctly.
- [x] If I changed the specification, I have checked that any validation
functions and tests reflect the changes.