Expose test fixture in pytest plugin (#135)
# Proposed Change
Expose example geff fixture in pytest plugin for use by other packages
with a geff dependency.

# Types of Changes
- Tests

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [ ] I have read the [developer/contributing](../CONTRIBUTING) docs.
- [ ] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [ ] I have checked that I maintained or improved code coverage.
- [ ] I have written docstrings and checked that they render correctly.
- [ ] If I changed the specification, I have checked that any validation
functions and tests reflect the changes.

---------

Co-authored-by: Draga Doncila <ddon0001@student.monash.edu>
Co-authored-by: Teun Huijben <45037215+TeunHuijben@users.noreply.github.com>