Specification updates: switch to an OME-like list of spatiotemporal axes with metadata (#147)
# Proposed Change
Continuing PR discussion started in #131

This PR is intended to close #80 and its subissues, but we need to go
through and check that all of the subissues are actually addressed.

- [x] #83 -- complete with the Axes basemodel 
- [x] #87 -- already closed by another PR
- [x] #102 -- already closed by another PR
- [x] #104 -- completed with this PR
- [x] #89 -- already closed by another PR

Remaining to dos:

- [ ] Compute min/max for axes props -> deferring on this until we have
the helper functions from #148
- [x] Validation/special handling for axes props, e.g. don't allow
missing nodes

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- New feature or enhancement
- Documentation update
- Tests

Which topics does your change affect? Delete those that do not apply.
- Specification
- `networkx` implementation

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the
[developer/contributing](https://github.com/live-image-tracking-tools/geff/blob/main/CONTRIBUTING)
docs.
- [ ] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [ ] I have checked that I maintained or improved code coverage.
- [x] I have written docstrings and checked that they render correctly.

## If you changed the specification
- [ ] I have checked that any validation functions and tests reflect the
changes.
- [ ] I have updated the GeffMetadata and the json schema using `pixi
run update-schema` if necessary.
- [ ] I have updated docs/specification.md to reflect the change.
- [ ] I have updated implementations to reflect the change. (This can
happen in separate PRs on a feature branch, but must be complete before
merging into main.)

# Further Comments
If this is a relatively large or complex change, kick off the discussion
by explaining why you chose the solution you did and what alternatives
you considered, etc...

---------

Co-authored-by: Caroline Malin-Mayor <malinmayorc@janelia.hhmi.org>
Co-authored-by: Teun Huijben <teun.huijben@czbiohub.org>
Co-authored-by: Teun Huijben <45037215+TeunHuijben@users.noreply.github.com>