Adding Python 3.13 (#142)
# Proposed Change
Adding missing Python 3.13 to tests and `pyproject.toml`.

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- Documentation update
- Maintenance (e.g. dependencies, CI, releases, etc.)

Which topics does your change affect? Delete those that do not apply.

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the [developer/contributing](../CONTRIBUTING) docs.
- [ ] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [ ] I have checked that I maintained or improved code coverage.
- [ ] I have written docstrings and checked that they render correctly.
- [ ] If I changed the specification, I have checked that any validation
functions and tests reflect the changes.

# Further Comments
If this is a relatively large or complex change, kick off the discussion
by explaining why you chose the solution you did and what alternatives
you considered, etc...

Co-authored-by: Caroline Malin-Mayor <malinmayorc@janelia.hhmi.org>