feat: FileReader class to read zarr to intermediate dict format (#129)
# Proposed Change
Adding a file reader class that allows a subset of the zarr to be read.
Follows a builder pattern, this allows each `node_prop` and `edge_prop`
to be read when required. When the `build` method is called, output
boolean masks can be provided to read only a subset of nodes and edges.

The `build` method outputs a new intermediate dictionary representation
of the graph defined as a `TypedDict`, this mirrors the structure in the
zarr file. This allows for reusability of the `FileReader` for different
backends.

A `read_to_dict` function has been added that outputs the intermediated
dict representation. The `read_nx` function now uses this and can also
accept a list of `node_props` and `edge_props` so that not all the
properties need to be read.

Tests have been added for the `FileReader`, it is also tested through
the `read_nx` texts.

# Types of Changes
- New feature or enhancement

Which topics does your change affect? Delete those that do not apply.
- `networkx` implementation

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the [developer/contributing](../CONTRIBUTING) docs.
- [ ] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [ ] I have checked that I maintained or improved code coverage.
- [x] I have written docstrings and checked that they render correctly.
- [ ] If I changed the specification, I have checked that any validation
functions and tests reflect the changes.

# Further Comments
If this is a relatively large or complex change, kick off the discussion
by explaining why you chose the solution you did and what alternatives
you considered, etc...

---------

Co-authored-by: melisande-c <melisande.croft@fht.org>
Co-authored-by: Morgan Schwartz <msschwartz21@gmail.com>
Co-authored-by: Caroline Malin-Mayor <malinmayorc@janelia.hhmi.org>