Refactor helper functions to provide write_array and write_dict (#148)
- Creates a helper for writing node/edge id arrays together, that checks
if the dtypes match
- Creates a helper for writing node or edge properties that are already
in array form
- Refactors the prior dictionary property writer to return the arrays
instead of writing them
- Uses the refactored dictionary helper in a write_dict_like function
that writes the whole graph to geff
- Uses this function in write_nx

Open todos and questions:
- [ ] test the helpers directly, not just through the nx implementation
- [x] is it too inefficient memory-wise to have the graph and all of the
properties attrs?
- [ ] Can update/add new attrs with the same helper function, I think,
but should be documented and tested
- [ ] We now enforce that node and edge ids must be the same dtype -
should we add this explicitly to the spec?

# Proposed Change
Briefly describe the contribution. If it resolves an issue or feature
request, be sure to link to that issue.

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- Bugfix (non-breaking change which fixes an issue)
- New feature or enhancement
- Documentation update
- Tests
- Maintenance (e.g. dependencies, CI, releases, etc.)

Which topics does your change affect? Delete those that do not apply.
- Specification
- `networkx` implementation

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the
[developer/contributing](https://github.com/live-image-tracking-tools/geff/blob/main/CONTRIBUTING)
docs.
- [ ] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [ ] I have checked that I maintained or improved code coverage.
- [ ] I have written docstrings and checked that they render correctly.

## If you have added or changed an implementation
- [ ] I wrote tests for the new implementation using standard fixtures
supplied in conftest.py.

# Further Comments
If this is a relatively large or complex change, kick off the discussion
by explaining why you chose the solution you did and what alternatives
you considered, etc...

---------

Co-authored-by: Teun Huijben <teun.huijben@czbiohub.org>
Co-authored-by: Teun Huijben <45037215+TeunHuijben@users.noreply.github.com>
Co-authored-by: Jan Funke <funkej@janelia.hhmi.org>