Fix docs styling and autogenerate html docs at mkdocs build-time (#168)
This PR updates documentation in a few ways:

1. fixes styles. The styles created by json-schema-for-humans was using
a relative path that was not valid when served by mkdocs... but they
also needed tweaking anyway. So this commits the styles and js to source
(in the docs folder) and tells mkdocs how to load them (in mkdocs.yml)
2. removes the need to run `generate-schema-doc` manually before running
`mike/mkdocs build/serve`. Now you can just checkout the repo and run
`uv run --group docs mkdocs serve` and it will autogenerate the schema
and place it where it needs to go.

---------

Co-authored-by: Morgan Schwartz <msschwartz21@gmail.com>