Add descriptions for each metadata field (#181)
# Proposed Change
Add descriptions to the `directed` and `axes` fields so they can be
displayed in the docs schema listing.

# Types of Changes
- Documentation update

Which topics does your change affect? Delete those that do not apply.
- Specification

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [ ] I have read the
[developer/contributing](https://github.com/live-image-tracking-tools/geff/blob/main/CONTRIBUTING)
docs.
- [ ] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [ ] I have checked that I maintained or improved code coverage.
- [ ] I have written docstrings and checked that they render correctly.

---------

Co-authored-by: Draga Doncila <ddon0001@student.monash.edu>
Co-authored-by: Morgan Schwartz <msschwartz21@gmail.com>