Add track-level node properties schema (#192)
# Proposed Change
Closes #151 

This PR adds a new, optional, `track_node_props` key to the geff
metadata that defines the node properties containing `tracklet` ID
and/or `lineage` ID information. It also explicitly defines what is
meant by a `tracklet` and a `lineage`.

# Types of Changes
- New feature or enhancement

Which topics does your change affect? Delete those that do not apply.
- Specification

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the
[developer/contributing](https://github.com/live-image-tracking-tools/geff/blob/main/CONTRIBUTING)
docs.
- [ ] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [ ] I have checked that I maintained or improved code coverage.
- [x] I have written docstrings and checked that they render correctly
by looking at the docs preview (link left as a comment on the PR).

## If you changed the specification
- [x] I have checked that any validation functions and tests reflect the
changes.
- [x] I have updated the GeffMetadata and the json schema using `pixi
run update-schema` if necessary.
- [x] I have updated docs/specification.md to reflect the change.
- [ ] I have updated implementations to reflect the change. (This can
happen in separate PRs on a feature branch, but must be complete before
merging into main.)

---------

Co-authored-by: Draga Doncila <ddon0001@student.monash.edu>