:bug: Remove duplicate ctc dependency group (#210)
# Proposed Change
Couldn't run pixi task update-json in the dev group because the ctc
group was included twice, once as a dependency group and once as extras.
Keep the extras, because we want users to install geff[ctc], similar to
spatial graph.

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- Maintenance (e.g. dependencies, CI, releases, etc.)