Support predefined object detection shapes (#159)
This adds storing object detections as pre-defined shapes on a geff graph ( #144).

We start off with
- circles and spheres as type `sphere` with a single number.
- 2D ellipses/3D ellipsoids as type `ellipsoid`, represented by a
2x2/3x3covariance matrix.

Thoughts:
- Convariance matrices: Store full matrix, using appropriate zarr compression for minimal overhead.

# Types of Changes
- New feature or enhancement

Which topics does your change affect? Delete those that do not apply.
- Specification
- Schema

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the
[developer/contributing](https://github.com/live-image-tracking-tools/geff/blob/main/CONTRIBUTING)
docs.
- [ ] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [x] I have checked that I maintained or improved code coverage.
- [x] I have written docstrings and checked that they render correctly.

## If you changed the specification
- [x] I have checked that any validation functions and tests reflect the
changes.
- [x] I have updated the GeffMetadata and the json schema using `pixi
run update-schema` if necessary.
- [x] I have updated docs/specification.md to reflect the change.
- [ ] I have updated implementations to reflect the change. (This can
happen in separate PRs on a feature branch, but must be complete before
merging into main.)

## If you have added or changed an implementation
- [ ] I wrote tests for the new implementation using standard fixtures
supplied in conftest.py.
- [ ] I updated pyproject.toml with new dependencies if needed.
- [ ] I added a function to tests/bench.py to benchmark the new
implementation.

---------

Co-authored-by: Jean-Yves Tinevez <tinevez@pasteur.fr>
Co-authored-by: Caroline Malin-Mayor <malinmayorc@janelia.hhmi.org>
Co-authored-by: Morgan Schwartz <msschwartz21@gmail.com>