Validator functions for geff data (#180)
# Proposed Change
This branch contains validators for checking different aspects of data
in geff in response to #152.
It also provides a generator of correct and incorrect geff graphs to be
replaced with a general approach in #161.

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- New feature or enhancement
- Tests

Which topics does your change affect? Delete those that do not apply.
- Specification
- `networkx` implementation

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the
[developer/contributing](https://github.com/live-image-tracking-tools/geff/blob/main/CONTRIBUTING)
docs.
- [x] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [x] I have checked that I maintained or improved code coverage.
- [ ] I have written docstrings and checked that they render correctly.

## If you changed the specification 
- no change

## If you have added or changed an implementation
- [ ] I wrote tests for the new implementation using standard fixtures
supplied in conftest.py.
- [ ] I updated pyproject.toml with new dependencies if needed.
- [ ] I added a function to tests/bench.py to benchmark the new
implementation.

---------

Co-authored-by: AnniekStok <anniek.stokkermans@gmail.com>
Co-authored-by: Caroline Malin-Mayor <malinmayorc@janelia.hhmi.org>