feat: Construct protocol and unified read function (#190)
# Proposed Change
Introduce a `ConstructFunc` protocol that takes series of NumPy arrays
that represent the GEFF data and constructs a graph in a specific
backend. If different backends follow this protocol they can be added to
the unified read function which will return a graph object based on the
`backend` flag. The supported backends are stored in an enum
`SupportedBackend`.

Added a readme with instructions on how to add read support for a new
backend.

Partially addresses #173.

### Testing

There is now a test for the `read` function that will be the same for
all new backends. All backends have to do is add the backend to the test
parametrisation and modify the `is_expected_type`, `get_nodes`,
`get_edges`, `get_node_prop` and `get_edge_prop` utility functions. The
test `test_read` duplicates `test_read_write_consistency`, but both
coexist.

Partially addresses #188

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- New feature or enhancement
- Tests

Which topics does your change affect? Delete those that do not apply.
- Implementation

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the
[developer/contributing](https://github.com/live-image-tracking-tools/geff/blob/main/CONTRIBUTING)
docs.
- [x] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [ ] I have checked that I maintained or improved code coverage.
- [ ] I have written docstrings and checked that they render correctly
by looking at the docs preview (link left as a comment on the PR).

## If you changed the specification
- [ ] I have checked that any validation functions and tests reflect the
changes.
- [ ] I have updated the GeffMetadata and the json schema using `pixi
run update-schema` if necessary.
- [ ] I have updated docs/specification.md to reflect the change.
- [ ] I have updated implementations to reflect the change. (This can
happen in separate PRs on a feature branch, but must be complete before
merging into main.)

## If you have added or changed an implementation
- [ ] I wrote tests for the new implementation using standard fixtures
supplied in conftest.py.
- [ ] I updated pyproject.toml with new dependencies if needed.
- [ ] I added a function to tests/bench.py to benchmark the new
implementation.

---------

Co-authored-by: melisande-c <melisande.croft@fht.org>