TrackMate XML to GEFF (#197)
# Proposed Change
Add a converter to convert TrackMate XML to GEFF.
For now, it only extracts the graph data, including nodes and edges
properties, and relies on write_nx to build a valid GEFF.
Relates to #69 .

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- New feature or enhancement
- Tests

Which topics does your change affect? Delete those that do not apply.
- io

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the
[developer/contributing](https://github.com/live-image-tracking-tools/geff/blob/main/CONTRIBUTING)
docs.
- [x] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [ ] I have checked that I maintained or improved code coverage.
- [ ] I have written docstrings and checked that they render correctly
by looking at the docs preview (link left as a comment on the PR).

# Further Comments
The actual implementation relies on iterparse from lxml (not part of
standard library). It means that the converter iteratively parses the
XML, line by line. At no point is the XML completely loaded into memory.
So, nice for crazy huge XMLs but probably comes at the cost of
efficiency.