Only check `path.exists` for local paths (#222)
# Proposed Change
The current validator precludes the handling of remote GEFFs because it
explicitly checks `path.exists`, which will always return `False` for a
remote GEFF.

This PR omits that check for URLs that are parsed as remote using the
builtin `urllib` library.

# Types of Changes
-  enhancement
- validators

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [ ] I have read the
[developer/contributing](https://github.com/live-image-tracking-tools/geff/blob/main/CONTRIBUTING)
docs.
- [x] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [x] I have checked that I maintained or improved code coverage.
- [ ] I have written docstrings and checked that they render correctly
by looking at the docs preview (link left as a comment on the PR).

---------

Co-authored-by: Draga Doncila <ddon0001@student.monash.edu>