rustworkx backend implementation (#94)
# Proposed Change
This PR adds `read_rx` and `write_rx` to support a rustworkx backend.
Closes #65 

# Types of Changes
- New feature or enhancement

Which topics does your change affect? Delete those that do not apply.
- `rustworkx` implementation

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the developer/contributing docs.
- [ ] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [ ] I have checked that I maintained or improved code coverage.
- [ ] I have written docstrings and checked that they render correctly.
- [ ] If I changed the specification, I have checked that any validation
functions and tests reflect the changes.

# Further Comments
For the `write_rx` I'm considering using the discussed higher level
`dict[str, Array]` writing API.

---------

Co-authored-by: melisande-c <melisande.croft@fht.org>