refactor: cleanup third-party imports and tests (#248)
closes #216 by cleaning up patterns around third party dependencies.
here's the overview:

- in `pyproject.toml`:
- `dev` environment is removed and made `default` (I think in most
cases, in local development, this is what you want. users just want to
`pixi install`, `pixi run` ... without having to use `-e dev` every
time. (cc @cmalinmayor and @msschwartz21 ... lemme know if you want that
reverted)
- removes duplicate third party features from the `dev` dependency
group.
- adds all the third party features manually to the default pixi
environment (because pixi doesn't support the standard
`"geff[spatial-graph,ctc,rx,lxml]"` syntax in dependency groups, so we
have to do it manually
- `ci.yaml`
- removes the internal reusable worfklow that was only being used in one
place (makes it easier to modify the test without creating one-off
inputs for each thing you want to modify)
- adds a test with *no* third-party extras, to ensure that there are no
assumptions about dependencies that "leak" into the package due to the
dev environment
- this also means that all tests must gracefully skip if their required
third-party package is missing (using
`pytest.skip(allow_module_level=True)`
- updates all third-party modules to check one-time at module import,
rather than each time inside a function. This pattern could be made
simpler if this package moved to private modules... but until then, we
have to assume that the end user is
"allowed" to import from anywhere. However, changing to all private
modules will cause lots of conflicts in open PRs and should be done
later
- simplified the imports in top level `geff.__init__` (here too, because
all modules are public, we can't assume that they will be importing from
the top, so we *have* to move those guards down to the lowest level
public export)
- drops the lxml dependency for trackmate. The standard library xml
module is sufficient for this use case and we should be able to support
trackmate without demanding an additional dependency. If the user *does*
install lxml, we can and do still use it (they have the same API), but
no need to fail without it. (cc @lxenard)
- I removed the `Attribute` type alias from trackmate_xml... there was a
place it still failed, and given how many things are in the union, I'm
not sure it's really adding that much value at this point. (cc @lxenard)
- updated type hints in trackmate_lxml (`ET.iterparse` is not valid as a
type hint, cc @lxenard)
- updates the segmentation type hint in `segmentation_validators` to
`ArrayLike`. No specific need to restrict it just to numpy, dask, and
zarr... there are plenty of other libraries that also implement an
arraylike api that we can easily support in this context (xarray, torch,
jax, etc...). cc @fjorka
- relaxes most of the `list[int] | tuple[int]` to `Sequence[int]`.
side-note: `tuple[int]` implies a tuple of length 1 a *single*
integer... `tuple[int, ...]` is what you want for a variable length
tuple of integers (cc @fjorka)
- adds `py.typed` to top level (otherwise mypy/pyright sees geff as
untyped ... even though it has a lot of type hints)

side-note: it would be nice not to bring in the dask dependency here
simply for the purpose of reading in a folder of tiffs and exporting to
zarr. tifffile itself can create a zarr store from a multi-page tiff or
folder of tiffs. (cc @JoOkuma)

apologies for all the things in a single PR. let me know if you want it
broken apart

---------

Co-authored-by: Morgan Schwartz <msschwartz21@gmail.com>
Co-authored-by: Laura Xénard <33941387+lxenard@users.noreply.github.com>