Add properties metadata to GeffMetadata specification (#253)
# Proposed Change
Contributes towards #244, but additional work is needed to close the
issue

Add the following optional fields to GeffMetadata:
- node_props_metadata
- edge_props_metadata
- tracklet_props_metadata
- lineage_props_metadata

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- New feature or enhancement

Which topics does your change affect? Delete those that do not apply.
- Specification

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the
[developer/contributing](https://github.com/live-image-tracking-tools/geff/blob/main/CONTRIBUTING)
docs.
- [ ] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [ ] I have checked that I maintained or improved code coverage.
- [x] I have written docstrings and checked that they render correctly
by looking at the docs preview (link left as a comment on the PR).

## If you changed the specification
- [ ] I have checked that any validation functions and tests reflect the
changes.
- [x] I have updated the GeffMetadata and the json schema using `pixi
run update-json` if necessary.
- [ ] I have updated docs/specification.md to reflect the change.
- [ ] I have updated implementations to reflect the change. (This can
happen in separate PRs on a feature branch, but must be complete before
merging into main.)

# Further Comments
No tests written for the new classes/functions for now => I will write
them in the other PR
But the current tests all pass (except spatial-graph because I don't
have the dependency)

---------

Co-authored-by: Morgan Schwartz <msschwartz21@gmail.com>