Reorganize cli to contain conversion functions in cli file (#243)
# Proposed Change
- Closes #220 
- Improves handling of imports for conversion functions which require
optional dependencies

Currently there still isn't any testing of raising errors when optional
imports are missing, but that's outside the scope of this PR.

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- Maintenance (e.g. dependencies, CI, releases, etc.)

Which topics does your change affect? Delete those that do not apply.

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the
[developer/contributing](https://github.com/live-image-tracking-tools/geff/blob/main/CONTRIBUTING)
docs.
- [x] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [ ] I have checked that I maintained or improved code coverage.
- [x] I have written docstrings and checked that they render correctly
by looking at the docs preview (link left as a comment on the PR).

---------

Co-authored-by: Teun Huijben <teun.huijben@czbiohub.org>
Co-authored-by: Teun Huijben <45037215+TeunHuijben@users.noreply.github.com>