remove pixi, add justfile (#264)
closes #261 

@msschwartz21, have a look and let me know how this feels for you.
Quickstart is:

```sh
brew install just
just --list
```

then start playing with commands.

One of the main difference from the pixi task runner is that you don't
have the intrinsic ability here to associate a task with a specific
environment... but uv makes it easy to ensure that task deps are
included by group. I didn't do it here, but if *don't* want the `uv run`
to update the environment in `.venv` you can always add `uv run
--isolated ...` to force it to use an ephemeral task-specific env