Properties metadata specification (#268)
# Proposed Change
Validation and tests for the new metadata specification

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- New feature or enhancement
- Tests

Which topics does your change affect? Delete those that do not apply.
- Specification

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the
[developer/contributing](https://github.com/live-image-tracking-tools/geff/blob/main/CONTRIBUTING)
docs.
- [x] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [x] I have checked that I maintained or improved code coverage.
- [ ] I have written docstrings and checked that they render correctly
by looking at the docs preview (link left as a comment on the PR).

## If you changed the specification
- [x] I have checked that any validation functions and tests reflect the
changes.
- [x] I have updated the GeffMetadata and the json schema using `pytest
--update-schema` if necessary.
- [ ] I have updated docs/specification.md to reflect the change.
- [ ] I have updated implementations to reflect the change. (This can
happen in separate PRs on a feature branch, but must be complete before
merging into main.)

# Further Comments
- For the dtype, I first tried to be more permissive by allowing
np.dtype | type | str, but it was a nightmare with pydantic and linting
so I went back to only allowing strings. I reused what was already done
for the validation of the arrays dtype.
- For the the encoding, since I have no clue what types of encodings
people may want to use, I tried to get the most encodings possible. But
there is no nice way to get a complete list of encodings and aliases
available in Python. The list of valid encodings I generate is already
quite long (~encodings!) but it doesn't include every aliases. Maybe
this is overkill and we should stick with just a few encodings?

---------

Co-authored-by: Morgan Schwartz <msschwartz21@gmail.com>