Update affine model to store numpy array directly, and provide schema hints (#269)
This PR updates the Affine model

closes #254 (in conjunction with #267)

The primary goal here is to:
1. provide better indication of what the matrix must be in the json
schema (i.e. list of list of float), which is done via the
`json_schema_input_schema` in the TFormMeta object.
2. get rid of `arbitrary_types_allowed` and store an actual numpy array
by giving pydantic all the info it needs to cast any object to a
well-formed square homogenous matrix (with `Annotated[np.ndarray,
TFormMeta]`)

cc @ilan-theodoro please review :)

-----

I stopped there, however I think it's worth discussing (in another PR)
whether there should be an `Affine` class *at all*. One could simply
make the GeffMetadata.affine field *directly* a `Annotated[np.ndarray,
TFormMeta]`... in which case the schema would look like:

```json
"affine": {
  "anyOf": [
    {
      "items": {
        "items": {
          "type": "number"
        },
        "type": "array"
      },
      "type": "array"
    },
    {
      "type": "null"
    }
  ],
  "default": null,
  "description": "Affine transformation matrix to transform the graph coordinates to the physical coordinates. The matrix must have the same number of dimensions as the number of axes in the graph."
},
```

This does mean you take some of the convenience methods off the
`Affine`... but I generally favor pure models that just say what they
are, without trying to add too much of a "convenience API" on top of it
(as tempting as it is! 😄 )... since it's just all the more to possibly
break. It doesn't stop you from having a functional module that adds
some of these conveniences, but it does makes the model clean and
direct. The main reason to *keep* the full `Affine` object is if you
intend to eventually have something *other* than one single `matrix`
field on it